+++ src/NovelSpider.Local.Jieqi/Properties/AssemblyInfo.cs
@@ -16,8 +16,8 @@ using System.Runtime.Versioning;
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyProduct("NovelSpider.Jieqi")]
 [assembly: AssemblyTitle("NovelSpider.Jieqi")]
-[assembly: AssemblyFileVersion("10.3.2.0")]
-[assembly: AssemblyVersion("10.3.2.0")]
+[assembly: AssemblyFileVersion("10.3.3.0")]
+[assembly: AssemblyVersion("10.3.3.0")]